add beginButton
