+++ src/components/newOrderButton.tsx
@@ -1,10 +1,11 @@
 import React from "react";
 import { AppButton } from "./appButton";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import {
   resetState,
   updateCurrentScene,
 } from "../store/appControl/appControlSlice";
+import { showNewOrderButtonSelector } from "../store/appControl/selectors";
 
 export function NewOrderButton() {
   const dispatch = useDispatch();
@@ -12,5 +13,9 @@ export function NewOrderButton() {
     dispatch(resetState());
     dispatch(updateCurrentScene("CREATOR"));
   };
-  return <AppButton text="new order" handledFunction={handleClickNewOrder} />;
+
+  const showNewOrderButton = useSelector(showNewOrderButtonSelector);
+  return showNewOrderButton ? (
+    <AppButton text="new order" handledFunction={handleClickNewOrder} />
+  ) : null;
 }
